Update to packer-plugin-sdk v0.0.11 and regenerate HCL config (#43)
* Update to packer-plugin-sdk v0.0.11 and regenerate HCL config

The packer 1.7 release changed the protocol version to 5 (also, notably,
the plugin SDK didn't correctly format the protocol information as
expected until v0.0.11). Running the packer-sdk-migrator tool mostly
worked, but left a couple of issues with fields in the config marked as
unused. It appears that the changes to HCL and how mapstructure decodes
things requires that you annotated basic fields with mapstructure
struct tags that don't follow the default mapping.

* Pull template/interpolate package from the packer SDK repo

The template/interpolate package has been moved to the packer-plugin-sdk
repository. This fixes the tests after rebasing.