Update targetOs to target_os to fix packer errors
The targetOs version of the parameter wasn't usable by packer directly,
and I received an error attempting to use the merged version of the
changes to use the 1.7 packer release. Turns out that when I rebased
onto the windows support, I failed to directly test with the same issue
reported in issue 40
[here](https://github.com/YaleUniversity/packer-provisioner-goss/issues/40#issuecomment-792780370)

I fixed a couple of issues around struct tags when updating the
provisioner to support 1.7; I just happened to miss the new parameter
added with Windows support.